+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file. This projec
 
 ## Unreleased
 
+### Added
+- [#2](https://github.com/laravel-json-api/encoder-neomerx/pull/2)
+New `withRequest` method on the encoder, allows a request to be passed into the encoding process.
+This is then provided to each `JsonApiResource` object when it is being encoded. This change
+was made to bring the resource into line with Laravel's Eloquent resource, which is passed the
+request when serializing models. *(Changes to classes in the `Schema` namespace to implement this
+are considered non-breaking, as they are internal classes.)*
+- Resource meta can now contain conditional attributes, as the conditional iterator is used when
+processing meta returned by the `JsonApiResource` class.
+
 ### Fixed
 - [#3](https://github.com/laravel-json-api/encoder-neomerx/issues/3)
 Add missing package discovery configuration to `composer.json`.